Refactor booking flow by consolidating the booking form into a reusable component with internationalization support. Split authentication logic into dedicated routes for login and registration, enhancing clarity and maintainability. Update API routes to utilize a single Prisma instance for consistent data access. Simplify the main booking page to render the new BookingForm component, improving the overall user experience.
