+++ app/api/auth/route.ts
@@ -61,99 +61,8 @@ const registerSchema = z.object({
  *         description: Validation error
  */
 export async function POST(request: NextRequest) {
-  try {
-    const body = await request.json();
-    const { pathname } = new URL(request.url);
-
-    if (pathname.endsWith('/login')) {
-      const { email, password } = loginSchema.parse(body);
-
-      const user = await prisma.user.findUnique({
-        where: { email },
-        select: {
-          id: true,
-          email: true,
-          name: true,
-          role: true,
-          password: true,
-        },
-      });
-
-      if (!user || !await bcrypt.compare(password, user.password)) {
-        return NextResponse.json(
-          { error: 'Invalid credentials' },
-          { status: 401 }
-        );
-      }
-
-      const token = jwt.sign(
-        { userId: user.id, email: user.email, role: user.role },
-        process.env.JWT_SECRET || 'fallback-secret',
-        { expiresIn: '7d' }
-      );
-
-      await prisma.session.create({
-        data: {
-          userId: user.id,
-          token,
-          expiresAt: new Date(Date.now() + 7 * 24 * 60 * 60 * 1000),
-        },
-      });
-
-      const userSanitized = { id: user.id, email: user.email, name: user.name, role: user.role };
-      return NextResponse.json({ token, user: userSanitized });
-    }
-
-    if (pathname.endsWith('/register')) {
-      const { email, password, name } = registerSchema.parse(body);
-
-      const existingUser = await prisma.user.findUnique({ where: { email } });
-      if (existingUser) {
-        return NextResponse.json(
-          { error: 'User already exists' },
-          { status: 400 }
-        );
-      }
-
-      const hashedPassword = await bcrypt.hash(password, 12);
-
-      const user = await prisma.user.create({
-        data: { email, password: hashedPassword, name },
-        select: { id: true, email: true, name: true, role: true },
-      });
-
-      const token = jwt.sign(
-        { userId: user.id, email: user.email, role: user.role },
-        process.env.JWT_SECRET || 'fallback-secret',
-        { expiresIn: '7d' }
-      );
-
-      await prisma.session.create({
-        data: {
-          userId: user.id,
-          token,
-          expiresAt: new Date(Date.now() + 7 * 24 * 60 * 60 * 1000),
-        },
-      });
-
-      return NextResponse.json({ token, user });
-    }
-
-    return NextResponse.json(
-      { error: 'Invalid endpoint' },
-      { status: 404 }
-    );
-  } catch (error) {
-    if (error instanceof z.ZodError) {
-      return NextResponse.json(
-        { error: 'Validation error', details: error.errors },
-        { status: 400 }
-      );
-    }
-    console.error('Auth error:', error);
-    return NextResponse.json(
-      { error: 'Internal server error' },
-      { status: 500 }
-    );
-  }
+  return NextResponse.json(
+    { error: 'Invalid endpoint. Use /api/auth/login or /api/auth/register' },
+    { status: 404 }
+  );
 }
