+++ docs/archive/finish-studio-space-mvp-phases-1-2-20250812.md
@@ -1,0 +1,63 @@
+# Task Archive: Finish Studio Space MVP (Phases 1-2)
+
+## Metadata
+- **Complexity**: Level 3
+- **Type**: Feature/System Refactor
+- **Date Completed**: $(date -u +%Y-%m-%dT%H:%M:%SZ)
+- **Related Docs**:
+  - `memory-bank/reflection.md`
+  - `memory-bank/tasks.md`
+  - Previous: `docs/archive/finish-studio-space-mvp-20250812.md`
+
+## Summary
+Phases 1-2 delivered security/data consolidation and a unified booking flow:
+- Single Prisma client via `@/lib/prisma` across APIs and `lib/auth.ts`.
+- Auth routes split into `/api/auth/login` and `/api/auth/register`.
+- RBAC enforced on admin endpoints.
+- Booking page now reuses `components/booking-form` with i18n and availability integration.
+
+## Requirements
+- Unify Prisma usage with a singleton wrapper.
+- Provide dedicated auth endpoints and persist sessions.
+- Enforce RBAC for admin APIs.
+- Unify booking UI and internationalize copy and navigation.
+
+## Implementation
+### Approach
+- Centralized data access in `lib/prisma` and refactored dependents.
+- Moved multiplexed auth logic into separate route files with Zod validation.
+- Leveraged existing `requireRole` for admin endpoints; validated consistency.
+- Consolidated booking experience into one reusable, localized client component.
+
+### Key Components
+- Data/Auth: `lib/prisma.ts`, `lib/auth.ts`
+- Auth API: `app/api/auth/login/route.ts`, `app/api/auth/register/route.ts`, `app/api/auth/route.ts`
+- Admin API: `app/api/admin/*`
+- Bookings API: `app/api/bookings/*`
+- Booking UI: `components/booking-form.tsx`, `app/[locale]/book/page.tsx`
+
+### Files Changed (high-level)
+- `lib/auth.ts`: switched to shared `prisma`.
+- `app/api/auth/login/route.ts`: new, login handler.
+- `app/api/auth/register/route.ts`: new, register handler.
+- `app/api/auth/route.ts`: simplified to 404.
+- `app/api/content/route.ts`: switched to shared `prisma`.
+- `components/booking-form.tsx`: i18n + locale-aware router; unified copy.
+- `app/[locale]/book/page.tsx`: now renders `BookingForm`.
+
+## Testing
+- Production build validated (Next 15). Known warning from `swagger-jsdoc` remains.
+- Manual verification of route registration in build output.
+
+## Lessons Learned
+- Consolidate the data layer first to reduce downstream churn.
+- Dedicated route files improve clarity and testing.
+- Using `@/i18n/routing` simplifies locale-aware navigation.
+
+## Future Considerations
+- Phase 3: Replace gallery images with `next/image`/`OptimizedImage` and correct sizes.
+- Phase 4: Add Jest + RTL smoke tests (auth, bookings, stats) and reevaluate CSP.
+
+## References
+- Reflection: `memory-bank/reflection.md`
+- Plan/Tasks: `memory-bank/tasks.md`
