+++ memory-bank/progress.md
@@ -3,3 +3,10 @@
 - Initialized Memory Bank via VAN mode.
 - Created core context files (brief, product, patterns, tech, active, tasks).
 - Archived interim state: 2025-08-11 -> docs/archive/finish-studio-space-mvp-20250812.md
+- VAN sync completed: 2025-08-12T09:27:50Z
+- PLAN verification complete: plan refined (included lib/auth.ts), Level 3 confirmed: 2025-08-12T09:31:20Z
+- IMPLEMENT Phase 1: prisma unified; auth routes split; build OK (Next 15) — 2025-08-12T09:42:22Z
+  - Files: lib/auth.ts, app/api/auth/login/route.ts, app/api/auth/register/route.ts, app/api/content/route.ts
+  - Build: success with swagger-jsdoc warning (known)
+- REFLECT: Phase 1-2 reviewed; reflection.md updated; checkpoints marked — 2025-08-12T09:56:23Z
+- ARCHIVE: Created docs/archive/finish-studio-space-mvp-phases-1-2-20250812.md; tasks marked COMPLETE — 2025-08-12T10:05:23Z
