+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const c=(c,i)=>(c=new URL(c+".js",i).href,s[c]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=s,document.head.appendChild(e)}else e=c,importScripts(c),s()})).then((()=>{let e=s[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e})));self.define=(i,t)=>{const a=e||("document"in self?document.currentScript.src:"")||location.href;if(s[a])return;let n={};const o=e=>c(e,a),r={module:{uri:a},exports:n,require:o};s[a]=Promise.all(i.map((e=>r[e]||o(e)))).then((e=>(t(...e),n)))}}define(["./workbox-5f3e40ee"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"6d53ca3fa54dd50e8cea1740b7560f5b"},{url:"/_next/static/MMtNzjGHx2dI5MfouAHLX/_buildManifest.js",revision:"6cfa76396137c173be0e0dcbe42eb629"},{url:"/_next/static/MMtNzjGHx2dI5MfouAHLX/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/160-4cdd2c75e4380bb2.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/191-5d46333c1da8891b.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/245-9518514735e7c215.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/306-4372aa726c567202.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/319-956ab293dc1396fe.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/408-79d1e24db8f89bdb.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/453-48eec8374a46dd1e.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/4bd1b696-88c914d9989ad21f.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/684-a9e35862ac8d8cb8.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/69-cd98ee1c9f10b9fd.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/776-da6dcde8614e14f4.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/812-6ec6cd4ce6cd0399.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/835-d97af165ebe89daf.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/admin/page-6ae509c686969fa0.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/api/admin/bookings/route-33beaae3f98e10cd.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/api/admin/stats/route-163138f1dc4ed4db.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/api/admin/users/route-3f128bb937b64348.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/book/page-c38560d2b7647dca.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/booking-success/page-263c184b0dc5313e.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/coworking/page-9a60e1611497e6bc.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/error-c1a59feb49580879.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/faq/page-d31f5312f7c88b34.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/gallery/page-1f7d26936dde84ba.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/layout-86f3e7fd939daa23.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/loading-60e8fb2d42cacf8c.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/not-found-d4a7ee36b529c4df.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/page-d185b3bf9ecfd66b.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/%5Blocale%5D/studio/page-486bd3d6608a7a7a.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/_not-found/page-c8335b631b2ba02c.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/api/admin/bookings/route-fc71298cf02a2ebe.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/api/admin/stats/route-223e8ff72ab1765a.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/api/admin/users/route-520f124720ce8537.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/api/analytics/vitals/route-3417c7ad777e1415.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/api/auth/route-57d7a4d488d1c438.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/api/bookings/availability/route-9ec32c6f0e181e2f.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/api/bookings/confirm/route-a7a6b57f014fae59.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/api/bookings/route-360f2271657d17e1.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/api/content/route-a73b3fa1e6587c05.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/app/api/docs/route-df9a1ad67747f484.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/framework-f593a28cde54158e.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/main-8035f2628c049960.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/main-app-21df8cfcdbb47447.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/pages/_app-da15c11dea942c36.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/pages/_error-cc3f077a18ea1793.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-c65d0d27869c13c9.js",revision:"MMtNzjGHx2dI5MfouAHLX"},{url:"/_next/static/css/23c07f7c5eb1f7c6.css",revision:"23c07f7c5eb1f7c6"},{url:"/_next/static/css/c66e46f533e6f7b8.css",revision:"c66e46f533e6f7b8"},{url:"/_next/static/media/26a46d62cd723877-s.woff2",revision:"befd9c0fdfa3d8a645d5f95717ed6420"},{url:"/_next/static/media/55c55f0601d81cf3-s.woff2",revision:"43828e14271c77b87e3ed582dbff9f74"},{url:"/_next/static/media/581909926a08bbc8-s.woff2",revision:"f0b86e7c24f455280b8df606b89af891"},{url:"/_next/static/media/8e9860b6e62d6359-s.woff2",revision:"01ba6c2a184b8cba08b0d57167664d75"},{url:"/_next/static/media/97e0cb1ae144a2a9-s.woff2",revision:"e360c61c5bd8d90639fd4503c829c2dc"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/_next/static/media/e4af272ccee01ff0-s.p.woff2",revision:"65850a373e258f1c897a2b3d75eb74de"},{url:"/browserconfig.xml",revision:"15191f398f9b3a84a131fe6c45c675d0"},{url:"/file.svg",revision:"d09f95206c3fa0bb9bd9fefabfd0ea71"},{url:"/globe.svg",revision:"2aaafa6a49b6563925fe440891e32717"},{url:"/icons/icon-128x128.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-144x144.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-152x152.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-192x192.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-384x384.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-512x512.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-72x72.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-96x96.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/logo.svg",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/manifest.json",revision:"3a489bf555db92a28c62d6eaa94db853"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/offline.html",revision:"62209e91aac27e78ea516ef292f02667"},{url:"/scripts/web-vitals.js",revision:"9ec08db201c72763fa5b612dacae8b3c"},{url:"/vasha_studio_logo.svg",revision:"20d112139a1f88b32ec1600c5d1ac103"},{url:"/vercel.svg",revision:"c0af2f507b369b085b35ef4bbe3bcf1e"},{url:"/window.svg",revision:"a2760511c65806022ad20adf74370ff3"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:c,state:i})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.googleapis\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.gstatic\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"images",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/^https:\/\/api\./i,new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((function(e){return"navigate"===e.request.mode}),new e.NetworkFirst({cacheName:"pages",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET")}));
+if(!self.define){let e,s={};const c=(c,i)=>(c=new URL(c+".js",i).href,s[c]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=s,document.head.appendChild(e)}else e=c,importScripts(c),s()})).then((()=>{let e=s[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e})));self.define=(i,a)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(s[n])return;let t={};const u=e=>c(e,n),r={module:{uri:n},exports:t,require:u};s[n]=Promise.all(i.map((e=>r[e]||u(e)))).then((e=>(a(...e),t)))}}define(["./workbox-5f3e40ee"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"3b0ecdc7d0aef4cef78013bf4b0b8efb"},{url:"/_next/static/P-S8ucGePYL2fuZM6hhs-/_buildManifest.js",revision:"38875a4f6244ab80a1a06256193b2b06"},{url:"/_next/static/P-S8ucGePYL2fuZM6hhs-/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/160-4cdd2c75e4380bb2.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/191-5d46333c1da8891b.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/199-2f8e99fa53f5e009.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/319-956ab293dc1396fe.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/359-78a09b0bd88640c3.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/408-79d1e24db8f89bdb.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/453-48eec8374a46dd1e.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/476-1916c44808d77073.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/4bd1b696-88c914d9989ad21f.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/671-f6856a0ac06edc6e.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/684-a9e35862ac8d8cb8.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/719-3b556eeb3934ddae.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/860-ddf3c6e2e6524cfc.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/admin/page-aea8614f7109dc26.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/api/admin/bookings/route-4d9309dd53c2959e.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/api/admin/stats/route-aa25e897614f977c.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/api/admin/users/route-e65afd8ad70e370d.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/book/page-30715f9dfa3ea177.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/booking-success/page-263c184b0dc5313e.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/coworking/page-9a60e1611497e6bc.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/error-c1a59feb49580879.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/faq/page-fd8edbdffe01f467.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/gallery/page-1f7d26936dde84ba.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/layout-db74e50dd5daf4e0.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/loading-4e9405a9fe266345.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/not-found-d4a7ee36b529c4df.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/page-d185b3bf9ecfd66b.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/%5Blocale%5D/studio/page-486bd3d6608a7a7a.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/_not-found/page-c8335b631b2ba02c.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/admin/bookings/route-34d0fcacad396366.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/admin/stats/route-5ec95b980df101af.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/admin/users/route-5466af5518b7356b.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/analytics/vitals/route-80fa746ca2e8869d.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/auth/login/route-1290347482cb04e4.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/auth/register/route-fe294cceb8c690e7.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/auth/route-07e8cf8940406df1.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/bookings/availability/route-16ea9a3245582a7d.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/bookings/confirm/route-b64b9cdd5b4844a1.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/bookings/route-45793c994070bfbc.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/content/route-4a6daac9051a46af.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/app/api/docs/route-29982170d001997e.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/framework-f593a28cde54158e.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/main-8035f2628c049960.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/main-app-21df8cfcdbb47447.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/pages/_app-da15c11dea942c36.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/pages/_error-cc3f077a18ea1793.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-c65d0d27869c13c9.js",revision:"P-S8ucGePYL2fuZM6hhs-"},{url:"/_next/static/css/69c4d2d228c65daf.css",revision:"69c4d2d228c65daf"},{url:"/_next/static/css/c66e46f533e6f7b8.css",revision:"c66e46f533e6f7b8"},{url:"/_next/static/media/26a46d62cd723877-s.woff2",revision:"befd9c0fdfa3d8a645d5f95717ed6420"},{url:"/_next/static/media/55c55f0601d81cf3-s.woff2",revision:"43828e14271c77b87e3ed582dbff9f74"},{url:"/_next/static/media/581909926a08bbc8-s.woff2",revision:"f0b86e7c24f455280b8df606b89af891"},{url:"/_next/static/media/8e9860b6e62d6359-s.woff2",revision:"01ba6c2a184b8cba08b0d57167664d75"},{url:"/_next/static/media/97e0cb1ae144a2a9-s.woff2",revision:"e360c61c5bd8d90639fd4503c829c2dc"},{url:"/_next/static/media/df0a9ae256c0569c-s.woff2",revision:"d54db44de5ccb18886ece2fda72bdfe0"},{url:"/_next/static/media/e4af272ccee01ff0-s.p.woff2",revision:"65850a373e258f1c897a2b3d75eb74de"},{url:"/browserconfig.xml",revision:"15191f398f9b3a84a131fe6c45c675d0"},{url:"/file.svg",revision:"d09f95206c3fa0bb9bd9fefabfd0ea71"},{url:"/globe.svg",revision:"2aaafa6a49b6563925fe440891e32717"},{url:"/icons/icon-128x128.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-144x144.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-152x152.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-192x192.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-384x384.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-512x512.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-72x72.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/icon-96x96.png",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/icons/logo.svg",revision:"9d6c8fa6ad6c56ec6840c5237ec846ec"},{url:"/manifest.json",revision:"3a489bf555db92a28c62d6eaa94db853"},{url:"/next.svg",revision:"8e061864f388b47f33a1c3780831193e"},{url:"/offline.html",revision:"62209e91aac27e78ea516ef292f02667"},{url:"/scripts/web-vitals.js",revision:"9ec08db201c72763fa5b612dacae8b3c"},{url:"/vasha_studio_logo.svg",revision:"20d112139a1f88b32ec1600c5d1ac103"},{url:"/vercel.svg",revision:"c0af2f507b369b085b35ef4bbe3bcf1e"},{url:"/window.svg",revision:"a2760511c65806022ad20adf74370ff3"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:c,state:i})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.googleapis\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.gstatic\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"images",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/^https:\/\/api\./i,new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((function(e){return"navigate"===e.request.mode}),new e.NetworkFirst({cacheName:"pages",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET")}));
